extract controller init to posmodel
